02-agent-survey/02-llamaindex/requirements.txt
llama-index-core==0.12.5
llama-index-llms-groq==0.3.1
python-dotenv==1.0.1
